+++ CHANGELOG.md
@@ -32,3 +32,4 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Maintenance
 ### Refactoring
 - Improve error messages for workflow states other than NOT_STARTED ([#642](https://github.com/opensearch-project/flow-framework/pull/642))
+- Create a Config XContent model for Config index ([#679](https://github.com/opensearch-project/flow-framework/pull/679))